segfaut audio linux

+++ includes/doom.h
@@ -217,7 +217,7 @@ t_game					player_properties(t_game game, t_event events);
 void					play_music(t_game game, size_t id, size_t vol,
 						size_t frame);
 t_game					animate(t_game game, float old_timer);
-t_array					stack_sounds(t_array chunk, size_t id, u_int32_t vol);
+t_array					stack_sounds(t_array chunk, size_t id, float vol);
 t_game					play_sounds(t_game game);
 
 void					render_multi_threaded(const t_env env, t_color *buf);
